wip: commit progress on test update
